undeprecate terminal.interactiveshell
move terminal.ptshell to terminal.interactiveshell, since it is the new implementation of the class.

The deprecation was there while the readline implementation remained,
but there is no need for it now that there is just one implementation of TerminalInteractiveShell.
